Deployed-Django/backend_logic/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.24.2
ortools==9.6.2534
protobuf==4.22.1
psycopg2-binary==2.9.5
pytz==2023.3
scipy==1.10.1
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
